+++ server/src/server.js
@@ -1,5 +1,5 @@
 const proxyServer = require('./proxy');
-const uploadServer = require('./upload');
+const internalAPI = require('./internalAPI');
 
 const proxyPort = 3000;
 const uploadPort = 5000;
@@ -8,6 +8,6 @@ proxyServer.listen(proxyPort, () => {
   console.log(`Proxy server listening at http://localhost:${proxyPort}`);
 });
 
-uploadServer.listen(uploadPort, () => {
+internalAPI.listen(uploadPort, () => {
   console.log(`Upload server listening at http://localhost:${uploadPort}`);
 });
